+++ CJBItemSpawner/docs/release-notes.md
@@ -1,6 +1,10 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+
+* Added Support for [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098).
+
 ## 2.3.3
 Released 21 March 2024 for SMAPI 4.0.0 or later.
 
@@ -253,7 +257,7 @@ Released 05 April 2017 for SMAPI 1.9 or later.
 * Fixed tab selection broken in 1.7.
 
 ## 1.7
-Released 05 April 2017 for SMAPI 1.9 or later. 
+Released 05 April 2017 for SMAPI 1.9 or later.
 
 * Updated to SMAPI 1.9.
 * Now only shows item sell price if you have the CJB Show Item Sell Price mod installed.